Build: create script to generate ESM-compatible code
The code, written in AMD module syntax, lives in the "ui"
folder.
A new Grunt task is added ("esmify") that will process every
JavaScript source file in the "ui" folder and convert the AMD
imports into ESM imports.
The transpiled code will be stored in the "esm" folder.

This means that consumers of jquery-ui can choose whether they
want to use the ESM version or the AMD version.